CLI for turning on 3.8V for power supplies (#153)
* function for turning on 3.8V power supplies, with bit mask per supply
* CLI for turning on 3.8V power supplies, granularity per FPGA
* clean up use of old-style prototypes: Replace () with (void)
* change format of board id from decimal to hex in printout

+++ projects/cm_mcu/I2CSlaveTask.c
@@ -55,8 +55,8 @@ static uint8_t getSlaveData(uint8_t address)
     case 0x0U: // reserved
       value = testreg;
       break;
-    case 0x10U:                       // MCU temperature
-      value = getADCvalue(20) + 0.5f; // always valid
+    case 0x10U:                                // MCU temperature
+      value = (uint8_t)getADCvalue(20) + 0.5f; // always valid
       break;
     case 0x12U: // FPGA F2 temp
       value = (uint8_t)(local_fpga_f2 >= 0 ? fpga_args.pm_values[local_fpga_f2] : 0U);
